// packages/api/src/types.ts
export interface PostCastResponse {
	success: boolean;
	cast: {
		hash: string;
		author: {
			fid: number;
		};
	};
	text: string;
}